Black Lives Matter
Inclusive is better than exclusive: this is a core principal of
Kubernetes.

When people are systematically murdered, dehumanized, marginalized,
threatened, and made to feel unwelcome, it's our job to make it clear
that such things will not fly in our community, and must not be allowed
to fly in society.

Without inclusivity, there is no community, and without community, there
is no Kubernetes, there is no KubeBuilder.

---

Verbiage and general design taken from the main Kubernetes website
(https://github.com/kubernetes/website/pull/21505).